size_t?

+++ 0x12-more_singly_linked_lists/1-listint_len.c
@@ -8,7 +8,7 @@
 
 size_t listint_len(const listint_t *h)
 {
-	int num_elem = 0;
+	size_t num_elem = 0;
 
 	while (h != NULL)
 	{
